Seperate directional collision class, added additional comments
